Refactor: 전역 상수 한곳에 모으기
