+++ frontend/src/globals/constants.ts
@@ -1,0 +1,2 @@
+export const DEFAULT_PROFILE_IMAGE = '/images/default_profile_image.jpg';
+export const LOGO_IMAGE = '/images/logo_image.svg';
