+++ frontend/src/globals/descriptions.ts
