+++ frontend/src/pages/echo/index.tsx
@@ -1,7 +1,7 @@
 import Head from 'next/head';
 
 import Header from 'src/components/Header';
-import descriptions from 'src/utils/descriptions';
+import descriptions from 'src/globals/descriptions';
 
 function Echo() {
   return (
