+++ frontend/src/pages/users/[username]/settings.tsx
@@ -10,7 +10,7 @@ import { Row } from 'src/components/Grid';
 import { UserType } from 'src/types';
 
 import { Fetcher } from 'src/utils';
-import descriptions from 'src/utils/descriptions';
+import descriptions from 'src/globals/descriptions';
 
 import { Page } from 'src/styles';
